added dying system and respawn
if a player dies, DieAndGoToHell.cs spawns a tombstone (need to make), and sets the player to inactive, and enables the player ghost

if a player walks into the tombstone collider they can hold down a button to begin the respawn. There is a min respawn time and a max respawn time.  Health at respawn is a percentage of time held down over max res time